+++ test/ClaimCollectionOwnership.t.sol
@@ -20,6 +20,7 @@ import {JBPermissionIds} from "@bananapus/permission-ids-v6/src/JBPermissionIds.
 import {JBSuckerDeployerConfig} from "@bananapus/suckers-v6/src/structs/JBSuckerDeployerConfig.sol";
 
 import {CTDeployer} from "../src/CTDeployer.sol";
+import {IPermit2} from "@uniswap/permit2/src/interfaces/IPermit2.sol";
 import {CTPublisher} from "../src/CTPublisher.sol";
 import {ICTPublisher} from "../src/interfaces/ICTPublisher.sol";
 import {CTAllowedPost} from "../src/structs/CTAllowedPost.sol";
@@ -79,7 +80,9 @@ contract ClaimCollectionOwnershipTest is Test {
         );
 
         // Deploy publisher.
-        publisher = new CTPublisher(IJBDirectory(makeAddr("directory")), permissions, feeProjectId, address(0));
+        publisher = new CTPublisher(
+            IJBDirectory(makeAddr("directory")), permissions, feeProjectId, IPermit2(address(0)), address(0)
+        );
 
         // Deploy the CTDeployer.
         ctDeployer = new CTDeployer(