+++ test/regression/RegressionMetadataShadow.t.sol
@@ -17,6 +17,7 @@ import {JBAccountingContext} from "@bananapus/core-v6/src/structs/JBAccountingCo
 import {JBPermissionsData} from "@bananapus/core-v6/src/structs/JBPermissionsData.sol";
 import {JBSplit} from "@bananapus/core-v6/src/structs/JBSplit.sol";
 
+import {IPermit2} from "@uniswap/permit2/src/interfaces/IPermit2.sol";
 import {CTPublisher} from "../../src/CTPublisher.sol";
 import {CTAllowedPost} from "../../src/structs/CTAllowedPost.sol";
 import {CTPost} from "../../src/structs/CTPost.sol";
@@ -174,7 +175,8 @@ contract RegressionMetadataShadowTest is Test {
         MetadataNoopTerminal feeTerminal = new MetadataNoopTerminal();
         directory.setTerminals(IJBTerminal(address(projectTerminal)), IJBTerminal(address(feeTerminal)));
 
-        CTPublisher publisher = new CTPublisher(IJBDirectory(address(directory)), permissions, 1, address(0));
+        CTPublisher publisher =
+            new CTPublisher(IJBDirectory(address(directory)), permissions, 1, IPermit2(address(0)), address(0));
 
         CTAllowedPost[] memory allowedPosts = new CTAllowedPost[](1);
         allowedPosts[0] = CTAllowedPost({