+++ test/regression/RegressionPublishHookBoundary.t.sol
@@ -16,6 +16,7 @@ import {JBCurrencyIds} from "@bananapus/core-v6/src/libraries/JBCurrencyIds.sol"
 import {JBAccountingContext} from "@bananapus/core-v6/src/structs/JBAccountingContext.sol";
 import {JBPermissionsData} from "@bananapus/core-v6/src/structs/JBPermissionsData.sol";
 import {JBSplit} from "@bananapus/core-v6/src/structs/JBSplit.sol";
+import {IPermit2} from "@uniswap/permit2/src/interfaces/IPermit2.sol";
 
 import {CTPublisher} from "../../src/CTPublisher.sol";
 import {CTAllowedPost} from "../../src/structs/CTAllowedPost.sol";
@@ -183,7 +184,11 @@ contract RegressionPublishHookBoundaryTest is Test {
         directory.setTerminal(feeProjectId, IJBTerminal(address(feeTerminal)));
 
         CTPublisher publisher = new CTPublisher(
-            IJBDirectory(address(directory)), IJBPermissions(address(permissions)), feeProjectId, address(0)
+            IJBDirectory(address(directory)),
+            IJBPermissions(address(permissions)),
+            feeProjectId,
+            IPermit2(address(0)),
+            address(0)
         );
 
         CTAllowedPost[] memory allowedPosts = new CTAllowedPost[](1);