+++ test/regression/ZeroAddressFeeBeneficiary.t.sol
@@ -17,6 +17,7 @@ import {JBAccountingContext} from "@bananapus/core-v6/src/structs/JBAccountingCo
 import {JBPermissionsData} from "@bananapus/core-v6/src/structs/JBPermissionsData.sol";
 import {JBSplit} from "@bananapus/core-v6/src/structs/JBSplit.sol";
 
+import {IPermit2} from "@uniswap/permit2/src/interfaces/IPermit2.sol";
 import {CTPublisher} from "../../src/CTPublisher.sol";
 import {CTAllowedPost} from "../../src/structs/CTAllowedPost.sol";
 import {CTPost} from "../../src/structs/CTPost.sol";
@@ -162,7 +163,7 @@ contract ZeroAddressFeeBeneficiaryTest is Test {
         feeTerminal = new ZBAcceptingTerminal();
         projectTerminal.configure({store_: store, hook_: address(hook)});
         feeTerminal.configure({store_: store, hook_: address(hook)});
-        publisher = new CTPublisher(IJBDirectory(address(directory)), permissions, 1, address(0));
+        publisher = new CTPublisher(IJBDirectory(address(directory)), permissions, 1, IPermit2(address(0)), address(0));
 
         directory.setTerminals(address(projectTerminal), address(feeTerminal));